move ListSlideURLs to integration_util_test.go; Used only on testing now
